requirements.txt
python-dotenv
requests
tiktoken
openai
pandas
